+++ Paramount.Betterclassifieds.Business/Booking/AdBookingModel.cs
@@ -75,5 +75,6 @@ namespace Paramount.Betterclassifieds.Business.Booking
         public List<Enquiry> Enquiries { get; set; }
         public int[] Publications { get; set; }
         public string CategoryAdType { get; set; }
+        public string CategoryFontIcon { get; set; }
     }
 }