+++ Paramount.Betterclassifieds.DataService/Search/ClassifiedsSearchEntities.cs
@@ -1,0 +1,6 @@
+namespace Paramount.Betterclassifieds.DataService.Search
+{
+    partial class ClassifiedsSearchEntitiesDataContext
+    {
+    }
+}